refactor(ui): rename Workflows to Pipelines for naming consistency
The server-side has used 'pipeline' exclusively (pipelines table, pipeline_runs,
PipelineController, /api/pipelines, etc.) since the early phases. The client UI
still surfaced 'Workflows' as the nav label and exposed a /workflow route that
redirected to /pipelines. The components/ tree was split across two directories
(components/workflow and components/pipeline) for the same domain.

Changes:
- MainLayout nav label 'Workflows' -> 'Pipelines' (href was already /pipelines)
- Drop the /workflow legacy route + its redirect stub page (Workflow.tsx)
- Move 6 components from components/workflow/ to components/pipeline/
  (AgentChat, AgentNode, CodePreview, MultiAgentPipeline, SandboxConfig,
  StrategyConfig). git mv preserves rename history.
- Update PipelineDetail.tsx imports
- Update the back-button Link in PipelineRun.tsx to point at /pipelines
- Update test file reference paths in tests/unit/pipeline-ux.test.ts

Not renamed (intentionally — these are GitHub Actions terminology, not product naming):
- TriggerForm.tsx 'workflow_run' GitHub event option
- server/mcp-servers/github/* tool names like github_list_workflows
