+++ client/src/components/pipeline/AgentChat.tsx
