+++ client/src/components/pipeline/CodePreview.tsx
