+++ client/src/components/pipeline/MultiAgentPipeline.tsx
