+++ client/src/components/pipeline/StrategyConfig.tsx
