+++ client/src/pages/PipelineDetail.tsx
@@ -7,9 +7,9 @@ import { Card } from "@/components/ui/card";
 import { ArrowLeft, Play, Loader2, GitCompare, Network } from "lucide-react";
 import { Checkbox } from "@/components/ui/checkbox";
 import { useLocation } from "wouter";
-import MultiAgentPipeline from "@/components/workflow/MultiAgentPipeline";
-import AgentChat from "@/components/workflow/AgentChat";
-import CodePreview from "@/components/workflow/CodePreview";
+import MultiAgentPipeline from "@/components/pipeline/MultiAgentPipeline";
+import AgentChat from "@/components/pipeline/AgentChat";
+import CodePreview from "@/components/pipeline/CodePreview";
 import DAGEditor from "@/components/dag/DAGEditor";
 import { usePipeline, useStartRun, useRuns } from "@/hooks/use-pipeline";
 import { usePipelineDAG } from "@/hooks/use-dag";
